requirements.txt
# XDR_SOC_AIAGENT - Python Dependencies
# Install: pip install -r requirements.txt

# HTTP requests
requests>=2.28.0

# Cryptography for AES decryption
pycryptodome>=3.15.0
